fixed init_mail (OK)

+++ partools/mail/__init__.py
@@ -1,5 +1,6 @@
 import ssl
 import smtplib
+import os.path as p
 import win32com.client as win32
 from shutil import copytree
 
@@ -28,6 +29,7 @@ def gmail(mail_name,
     pwd = gl.cfi['PWD_GMAIL']
 
     ctx = ssl.create_default_context()
+    u.log(f"Sending mail '{mail_name}' to {gl.recipients}...")
     with smtplib.SMTP_SSL(host, port, context=ctx) as server:
         server.login(user, pwd)
         server.sendmail(gl.sender, gl.recipients, msg.as_string())
@@ -46,6 +48,7 @@ def no_auth(mail_name,
     init_cfi()
     msg = get.msg(subject, TXTbody, HTMLbody, attachments, var_dict)
 
+    u.log(f"Sending mail '{mail_name}' to {gl.recipients}...")
     with smtplib.SMTP(gl.NO_AUTH_HOST) as server:
         server.sendmail(gl.sender, gl.recipients, msg.as_string())
     u.log('Mail sent')
@@ -73,6 +76,7 @@ def outlook(mail_name,
     mail.HTMLBody = HTMLbody
     for attachment in attachments:
         mail.Attachments.Add(attachment)
+    u.log(f"Sending mail '{mail_name}' to {gl.recipients}...")
     mail.Send()
     u.log('Mail sent')
 
@@ -84,22 +88,26 @@ def init(mail_name, recipients, check_internal=False):
         gl.recipients = recipients
     else:
         gl.recipients = get.recipients(check_internal)
-    u.log(f"Sending mail '{mail_name}' to {gl.recipients}...")
 
 
 def init_cfi():
 
     gl.cfi = u.g.get_confidential(False)
     if not gl.cfi:
-        u.log(gl.S_MISSING_CFI)
-        raise Exception(u.g.E_CFI)
+        raise Exception(gl.S_MISSING_CFI)
     gl.sender = gl.cfi['MAIL_FROM']
     gl.From = gl.cfi['MAIL_FROM']
 
 
 def init_mail():
 
+    if p.exists(cfg.MAILS_DIR):
+        u.log(f"'{cfg.MAILS_DIR}' already exists."
+              " Initialisation of mail folder aborted.")
+        return
+    files_dir = f'{p.dirname(__file__)}/files'
+    files_dir = files_dir.replace('\\', '/')
     u.delete_folder(cfg.MAILS_DIR)
-    copytree('partools/test/mails', cfg.MAILS_DIR)
+    copytree(files_dir, cfg.MAILS_DIR)
     u.save_list(['*'], cfg.MAILS_DIR + '.gitignore')
     u.log(f"Mail folder '{cfg.MAILS_DIR}' successfully initialised")
